requirements.txt
requests
python-dotenv
flask
anthropic
yt-dlp
youtube-transcript-api